requirements.txt
# === Core App & UI ===
streamlit==1.33.0
python-dotenv==1.0.1

# === Whisper (OpenAI STT) ===
openai-whisper==20231117
ffmpeg-python==0.2.0
soundfile==0.12.1
numpy>=1.21
torch>=2.1.0
torchaudio>=2.1.0

# === Speaker Diarization ===
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.pipeline==3.0.1
huggingface-hub>=0.14.1

# === Transformers (Summarization, Translation, DeepSeek, etc.)
transformers==4.39.3
sentencepiece==0.1.99
accelerate==0.29.3

# === DeepSeek models (optional)
einops>=0.6.0

# === Audio recording
sounddevice==0.4.6

# === NLP Utilities
scikit-learn>=1.3.2
dateparser==1.2.0
nltk>=3.8.1

# === Optional Debugging/Logs
rich>=13.0.0
